[CHORE] 패키지 오타 수정
