+++ src/main/java/com/official/memento/schedule/controller/dto/request/ScheduleCreateRequest.java
@@ -1,12 +1,13 @@
-package com.official.memento.schedule.conntroller.dto.request;
+package com.official.memento.schedule.controller.dto.request;
 
 import com.official.memento.global.util.Validator;
 import io.swagger.v3.oas.annotations.media.Schema;
+
 import java.time.LocalDateTime;
 
-@Schema(name = "일정 수정 요청")
-public record ScheduleUpdateRequest(
-        @Schema(description = "일정 내용", example = "수정된 일정")
+@Schema(name = "단일 일정 생성 요청")
+public record ScheduleCreateRequest(
+        @Schema(description = "일정 내용", example = "일정디스크립션")
         String description,
         @Schema(description = "일정 시작 날짜", example = "2025-01-20T10:00:00")
         LocalDateTime startDate,
@@ -17,7 +18,7 @@ public record ScheduleUpdateRequest(
         @Schema(description = "태그 아이디", example = "1")
         long tagId
 ) {
-    public ScheduleUpdateRequest of(
+    public ScheduleCreateRequest(
             final String description,
             final LocalDateTime startDate,
             final LocalDateTime endDate,
@@ -25,7 +26,11 @@ public record ScheduleUpdateRequest(
             final long tagId
     ) {
         checkNullData(description, startDate, endDate);
-        return new ScheduleUpdateRequest(description, startDate, endDate, isAllDay, tagId);
+        this.description = description;
+        this.startDate = startDate;
+        this.endDate = endDate;
+        this.isAllDay = isAllDay;
+        this.tagId = tagId;
     }
 
     private static void checkNullData(final String description, final LocalDateTime startDate, final LocalDateTime endDate){
@@ -35,3 +40,4 @@ public record ScheduleUpdateRequest(
         Validator.validateLengthContainEmoji(description, 30);
     }
 }
+
