+++ src/main/java/com/official/memento/schedule/controller/dto/request/ScheduleUpdateRequest.java
@@ -1,15 +1,12 @@
-package com.official.memento.schedule.conntroller.dto.request;
+package com.official.memento.schedule.controller.dto.request;
 
-import com.official.memento.global.exception.InvalidRequestBodyException;
 import com.official.memento.global.util.Validator;
 import io.swagger.v3.oas.annotations.media.Schema;
-
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 
-@Schema(name = "단일 일정 생성 요청")
-public record ScheduleCreateRequest(
-        @Schema(description = "일정 내용", example = "일정디스크립션")
+@Schema(name = "일정 수정 요청")
+public record ScheduleUpdateRequest(
+        @Schema(description = "일정 내용", example = "수정된 일정")
         String description,
         @Schema(description = "일정 시작 날짜", example = "2025-01-20T10:00:00")
         LocalDateTime startDate,
@@ -20,7 +17,7 @@ public record ScheduleCreateRequest(
         @Schema(description = "태그 아이디", example = "1")
         long tagId
 ) {
-    public ScheduleCreateRequest(
+    public ScheduleUpdateRequest of(
             final String description,
             final LocalDateTime startDate,
             final LocalDateTime endDate,
@@ -28,11 +25,7 @@ public record ScheduleCreateRequest(
             final long tagId
     ) {
         checkNullData(description, startDate, endDate);
-        this.description = description;
-        this.startDate = startDate;
-        this.endDate = endDate;
-        this.isAllDay = isAllDay;
-        this.tagId = tagId;
+        return new ScheduleUpdateRequest(description, startDate, endDate, isAllDay, tagId);
     }
 
     private static void checkNullData(final String description, final LocalDateTime startDate, final LocalDateTime endDate){
@@ -42,4 +35,3 @@ public record ScheduleCreateRequest(
         Validator.validateLengthContainEmoji(description, 30);
     }
 }
-
